encapsulated exepctions
