+++ SimpleMem/Memory.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using SimpleMem.Exceptions;
+using System;
 using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
